wado-compiler: extract type_fixup submodule from cm_binding
Move post-synthesis type fixups and call-site rewriting (the
recursive WASI-derived TypeId replacement, the binding-body fixups,
the rewrite_calls_in_block / collect_effect_calls_in_block walkers)
into `synthesis/cm_binding/type_fixup.rs`.
